requirements.txt
flask==3.1.1
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
groq==0.25.0
Pillow==11.2.1
opencv-python-headless==4.11.0.86
pytz==2025.1
line-bot-sdk==3.5.1
twilio==9.5.1
google-auth==2.40.1
google-api-python-client==2.166.0
gunicorn==23.0.0
